+++ tests/File/ComposerTest.php
@@ -13,13 +13,14 @@ class ComposerTest extends AbstractFile
     {
         $instance = clone self::getBingGeneratorInstance();
         $instance
-            ->setOptionPrefix('Api');
+            ->setOptionPrefix('Api')
+            ->setOptionComposerName('wsdltophp/bing');
         $composerFile = new Composer($instance, 'composer');
         $composerFile
             ->setRunComposerUpdate(false)
             ->write();
 
-        $this->assertSameFileContent('ValidBingComposer', $composerFile, 'json');
+        $this->assertSameFileContent('ValidBingComposer' . (version_compare(PHP_VERSION, '5.4.0') === -1  ? '.php53' : ''), $composerFile, 'json');
     }
     /**
      *
@@ -40,8 +41,7 @@ class ComposerTest extends AbstractFile
     public function testGetFileName()
     {
         $instance = clone self::getBingGeneratorInstance();
-        $instance
-            ->setOptionPrefix('Api');
+        $instance->setOptionPrefix('Api');
         $composerFile = new Composer($instance, 'composer');
 
         $this->assertSame(self::getTestDirectory() . 'composer.json', $composerFile->getFileName());