+++ tests/Parser/Wsdl/TagImportTest.php
@@ -39,7 +39,7 @@ class TagImportTest extends WsdlParser
 
         $schemaContainer = new SchemaContainer($tagImportParser->getGenerator());
         for ($i=0; $i<19; $i++) {
-            $schemaPath = realpath(sprintf(__DIR__ . '/../../resources/PartnerService.%d.xsd', $i));
+            $schemaPath = realpath(sprintf(__DIR__ . '/../../resources/partner/PartnerService.%d.xsd', $i));
             $schema = new Schema($tagImportParser->getGenerator(), $schemaPath, file_get_contents($schemaPath));
             $schema->getContent()->setCurrentTag('import');
             $schemaContainer->add($schema);